refactor: validacoes movidas para servico validador
